LayoutTests/imported/w3c/web-platform-tests/cssom/medialist-interfaces-004-expected.txt


PASS appendMedium_correctly_appends_medium_to_empty_MediaList 
FAIL appendMedium_correctly_appends_medium_to_nonempty_MediaList assert_equals: expected "all, screen" but got "screen, all"
